Creating usersController
