+++ src/modules/users/infra/http/controllers/SessionsController.ts
@@ -1,0 +1,19 @@
+import { Request, Response } from 'express';
+import { container } from 'tsyringe';
+
+import AuthenticateUserService from '@modules/users/services/AuthenticateUserService';
+
+export default class SessionsController {
+  public async create(request: Request, response: Response): Promise<Response> {
+    const { email, password } = request.body;
+
+    const userAuthenticate = container.resolve(AuthenticateUserService);
+
+    const { user, token } = await userAuthenticate.execute({
+      email,
+      password,
+    });
+
+    return response.json({ user, token });
+  }
+}
