+++ src/modules/users/infra/http/routes/users.routes.ts
@@ -1,41 +1,25 @@
 import { Router } from 'express';
 import multer from 'multer';
-import { container } from 'tsyringe';
 
-import CreateUserService from '@modules/users/services/CreateUserService';
-import UpdateUserAvatarService from '@modules/users/services/UpdateUserAvatarService';
 import ensureAuthenticated from '@modules/users/infra/http/middlewares/ensureAuthenticated';
 import uploadConfig from '@config/upload';
 
+import UsersController from '../controllers/UsersController';
+import UserAvatarController from '../controllers/UserAvatarController';
+
 const usersRouter = Router();
+const usersController = new UsersController();
+const userAvatarController = new UserAvatarController();
+
 const upload = multer(uploadConfig);
 
-usersRouter.post('/', async (request, response) => {
-  const { name, email, password } = request.body;
-  const createUserService = container.resolve(CreateUserService);
-
-  const user = await createUserService.execute({
-    name,
-    email,
-    password,
-  });
-
-  return response.json(user);
-});
+usersRouter.post('/', usersController.create);
 
 usersRouter.patch(
   '/avatar',
   ensureAuthenticated,
   upload.single('avatar'),
-  async (request, response) => {
-    const updateUserAvatarService = container.resolve(UpdateUserAvatarService);
-
-    const user = await updateUserAvatarService.execute({
-      user_id: request.user.id,
-      avatarFilename: request.file.fieldname,
-    });
-    return response.json(user);
-  },
+  userAvatarController.update,
 );
 
 export default usersRouter;
